remove redundant tests from msg server and refactor overly complex ones
