+++ x/dex/keeper/integration_placelimitorder_test.go
@@ -407,3 +407,69 @@ func (s *MsgServerTestSuite) TestPlaceLimitOrderInsufficientFunds() {
 	err := sdkerrors.ErrInsufficientFunds
 	s.assertAliceLimitSellFails(err, "TokenA", 0, 10)
 }
+
+func (s *MsgServerTestSuite) TestLimitOrderPartialFillDepositCancel() {
+	s.fundAliceBalances(100, 100)
+	s.fundBobBalances(100, 100)
+	s.assertDexBalances(0, 0)
+
+	s.aliceLimitSells("TokenB", 0, 50)
+
+	s.assertAliceBalances(100, 50)
+	s.assertBobBalances(100, 100)
+	s.assertDexBalances(0, 50)
+	s.assertCurrentTicks(math.MinInt64, 0)
+	s.assertMaxTick(0)
+	s.assertMinTick(math.MaxInt64)
+
+	s.bobMarketSells("TokenA", 10, 10)
+
+	s.assertAliceBalances(100, 50)
+	s.assertBobBalances(90, 110)
+	s.assertDexBalances(10, 40)
+	s.assertCurrentTicks(math.MinInt64, 0)
+	s.assertMaxTick(0)
+	s.assertMinTick(math.MaxInt64)
+
+	s.aliceLimitSells("TokenB", 0, 50)
+
+	s.assertAliceBalances(100, 0)
+	s.assertBobBalances(90, 110)
+	s.assertDexBalances(10, 90)
+	s.assertCurrentTicks(math.MinInt64, 0)
+	s.assertMaxTick(0)
+	s.assertMinTick(math.MaxInt64)
+
+	s.aliceCancelsLimitSell("TokenB", 0, 0)
+
+	s.assertAliceBalances(100, 40)
+	s.assertBobBalances(90, 110)
+	s.assertDexBalances(10, 50)
+	s.assertCurrentTicks(math.MinInt64, 0)
+	s.assertMaxTick(0)
+	s.assertMinTick(math.MaxInt64)
+
+	s.bobMarketSells("TokenA", 10, 10)
+
+	s.assertAliceBalances(100, 40)
+	s.assertBobBalances(80, 120)
+	s.assertDexBalances(20, 40)
+
+	s.aliceCancelsLimitSell("TokenB", 0, 1)
+
+	s.assertAliceBalances(100, 80)
+	s.assertBobBalances(80, 120)
+	s.assertDexBalances(20, 0)
+
+	s.aliceWithdrawsLimitSell("TokenB", 0, 0)
+
+	s.assertAliceBalances(110, 80)
+	s.assertBobBalances(80, 120)
+	s.assertDexBalances(10, 0)
+
+	s.aliceWithdrawsLimitSell("TokenB", 0, 1)
+
+	s.assertAliceBalances(120, 80)
+	s.assertBobBalances(80, 120)
+	s.assertDexBalances(0, 0)
+}
